Fix a bunch of bugs, break apart dataset dto
