+++ docker-compose.yml
@@ -7,8 +7,8 @@ services:
       POSTGRES_USER: postgres
       POSTGRES_PASSWORD: postgres
       POSTGRES_DB: statswales-backend
-    volumes:
-      - .docker/postgres/data:/var/lib/postgresql/data
+    #    volumes:
+#      - .docker/postgres/data:/var/lib/postgresql/data
 
   db-test:
     image: postgres:latest
