+++ package.json
@@ -12,7 +12,8 @@
     "lint:ci": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
     "build": "npm-run-all clean tsc copy-assets",
-    "pretest": "docker compose up -d db-test",
+    "pretest-docker": "docker compose up -d db-test",
+    "pretest-podman": "podman compose up -d db-test",
     "test": "NODE_ENV=test jest --coverage",
     "test:ci": "NODE_ENV=test jest --ci --coverage --config=jest.config.ts",
     "check": "npm-run-all prettier:fix lint:fix test build",
